backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dateutil>=2.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
openpyxl>=3.1.0

